Improve error message for uninitialized attribute
With the improved check for uninitialized class attributes, I found
multiple latent bugs in other Acton projects but as the error pointed to
the definition of the attribute and not where it is missing, it was
quite tedious to find where I had to apply the fix. For example, if
class Foo defines attribute x and then class Bar inherits from Foo, the
error message pointed to the parent class Foo's attribute x but the fix
is obviously in Foo.__init__ where we need to initialize self.x

The new error message includes "all" the information, for example:

    [error Type error]: Attribute 'y' is not initialized in TestClass.__init__
         +--> uninit_grandparent2.act@14:5-17:1
         |
       6 |    class BaseB(BaseA):
         :          ^------------
         :          `- Attribute inherited from BaseB
       7 |        y: int
         :        ^
         :        `- Attribute 'y' is defined here
         :
      13 |    class TestClass(BaseB):
         :    ^----------------------
         :    `- In class TestClass
      14 | +>     def __init__(self):
      15 | |          BaseA.__init__(self)
      16 | |          # Should get a failure since y is not initialized
      17 | |>
         : |
         : `- Attribute 'y' is not initialized in __init__
    -----+

It might even be a little too verbose. The parent class and class lines
are included as context and wouldn't actually need a text message
attached to them, but the Diagnose library we use to print error message
has no such facility, so we have to attach a message for now.