Only pass build-related args to actonc
When building the cmd for invoking actonc from the acton CLI, we have
used a sort of pass-through mode where most arguments are automatically
propagated. It made sense to begin with and with most arg being
build-related, however, more and more arguments are not build-related
and so it makes more sense to flip the logic. We now explicitly lists
the args to propagate to actonc when building!