Support min/max time / iterations for testing
We now have multiple parameters to control how many test iterations we
normally run and rather than using static values we focus on using
ranges by setting min and max values. The default is to do at least 3
iterations and run for 50ms. Most small unit tests functions run in
microseconds, so we can proably achieve hundreds of runs in 50ms. A
larger function (compute Pi, fibonacci etc) or something doing IO might
run for 100ms, surpassing the 50ms, but we allow that since we want to
have at least 3 runs (check flakiness). However, we have max values as
well, so we try to put an upper cap - if a test took 5 seconds to run,
we will stop after just 1 iteration.

In addition to max/min values, we can also set an exact number of
iterations, which under the hood just maps to using the same value for
--min-iter & --max-iter and setting a permissive time range.