Remove numpy support from compiler
The compiler has had special built-in support for numpy since it works
so differently in so many different ways. We are now removing it because
numpy remains unused and we want to avoid the extra maintenance burden.
We might very well bring numpy back at a later point. The later the
better since we are likely to improve other things in the compiler
meanwhile which would make a reintroduction easier.