Remove stub compilation mode
Stub mode was only used for numpy module which has been removed. The new
style is to use .ext.c files and NotImplemented in the body of functions
rather than the legacy stub mode compilation, so even if we decide to
bring back numpy, it will never be using stub mode again. This
simplifies the compilation pipeline.