Allow > 3.0 s test time(out)
The test timeout has been hardcoded to 3.0 seconds which is quite
obviously bad since if we want to run a test for a minimum time of more
than 3 seconds, that would always lead to a test timeout. Also, we now
automatically bump up max time based on the min time.