Add Acton LLDB plugin
This is a plugin for LLDB that helps decode Acton values, demangling function
names, argument names etc.

There are acton specific commands, so loading this in a plain LLDB, you can do
`acton bt` to print an Acton backtrace while the "real" C backtrace is still
available via `bt`.

We do hook in so that through LLDB-DAP, the normal backtrace is now the `acton
bt` and similarly for showing frame variables