Emit #line directives for sane backtraces
We now emit #line directives in CodeGen so that the resulting binaries we
produce are annotated with debug information on source location that points to
Acton locations rather than the C location. We use the .act filename rather
than the .c file as well.

There is a --dbg-no-lines flag to disable #line emission in case we need to
debug our C code.