Load .ty instead of always parsing .act
Previously we used to always parse the .act files even though we later
on would determine the .ty file is up to date and thus load that rather
than compile the .act. Now, we move the up-to-date check to the head of
the pipeline and if the .ty is up-to-date, we load it. We already store
imports in the .ty file and the typed module was recently added so we
get everything we need from the .ty file. To differentiate between
modules that we actually need to compile, we add a new TyTask which is
just about loading the content of a .ty file rather than running actual
compiler passes.

Parsing is usually a few milliseconds, sometimes tens of ms, but for
large modules it takes seconds, so being able to instead load the cached
.ty file speeds things up quite a lot.