Speed up builds with efficient .ty header loading
We now store a bit more information in the .ty file and rearrange the
fields so that it is possible to read these "header" fields quickly and
efficiently, from the start up to end of headers. Freshness checks,
import chasing and root generation now use the header to avoid decoding
large sections. This reduces I/O and decoding on no-op and partial
builds, making rebuilds faster.