Only regen doc index on project build
compilesFiles might be invoked for a small set of files, perhaps a
single file, but clearly generating an index for a project is only valid
for when we are working with the whole project, thus moving it to
buildProject.

It is a whole that the user could invoke `acton doc` before the project
is built, but that is not a new problem, it existed before.