Avoid rebuild by hashing public type of deps
We now compute the content hash of the public interface of a module and
store this in the .ty file, next to the normal content hash. The
difference is that the public interface only contains type signatures of
top level functions, constants and classes (with their methods etc).
Dependent modules do not need to be recompiled if the public interface
remains the same and we now make use of this fact. The imports field of
.ty files now tracks not just the module being imported but also which
content hash. Along with it, the dependency graph staleness algorithm
has been modified. Previously, a module would be considered stale if any
of its imported module were stale. Now, we also consider the public
interface hash of the imported modules - only if it changes does the
staleness propagate. Thus, if an imported module's public interface is
unchanged, we do not have to recompile our module - this is a tremendous
gain!!!

We bump the .ty version to 0.5 to force rebuilds. There's a new test
case to cover this as well.