Consider searchpath for .ty staleness
I made a crucial mistake in my implementation of reading iface hashes
from .ty header files of imports - if we couldn't find the .ty file,
we'd consider it "not changed" - how wreckless of me. If we cannot find
the .ty file, clearly that should be an error... and now we actually go
through the search path looking for the .ty file, like we do in other
places where we access .ty files.

I basically wrote the .ty freshness check based on projects without any
dependencies. Now we properly support project deps!