Enable actonc GHC parallel (big) nursery
We enable parallel GC, which acts on the first generation 0 - the
nursery, and set the size of the nursery to 32MB. This improves
performance for our normal compilation tasks, something like 10%, with
no extra effort.

The actonc Haskell GHC runtime system is now threaded but there are
likely no other drawbacks to that. It's been debated for years whether
GHC should enable a threaded runtime per default and the argument
against it is that in Haskell, there are programs that run slower in a
threaded environment, essentially because they are single threaded and
the no-threads environment avoids some overhead. In our case, the
threaded runtime is provably faster, so it's a simple choice for us.
This is faster, let's go!