Concurrently build modules
actonc is now threaded - well, that was recently enabled separately, to
get parallel GC - but now we actually make use of it by compiling
multiple modules concurrently using multiple threads. Scaleout isn't
perfect, like we are seeing something like a 10-20% drop in per-module
performance. I.e., a module that took 10 seconds to compile before now
takes like 11-12 seconds, but in exchange we can actually compile
modules concurrently so the overall throughput is close to linear
improvement, i.e. 10 modules compile like 9x as fast now, given there
are at least 10 CPU cores available.

The default is to run as many concurrent modules as there are CPU cores
available.

The work split and data flow is actor-like, in that there is a central
main thread that coordinates the work. It maintains the Env that
contains compiled modules. When there is work to compile a module, it
creates async thread workers to compile each module, feeding in the
shared env at that point - like a snapshot. Once a module compilation
worker is done, the result is fed back to the main coordinating thread,
which extends the environment and starts new worker threads for the next
tasks. This means we avoid sharing any state between threads with MVar
or similar and overall get a simpler and more readable control flow.

We also have some hackery for doing a global cache for loading .ty
files, using MVars to get fast cached reads.