Prioritize compilation order by critical-path
Instead of picking the module to compile randomly in a project, we pick
by the critical-path. That is, we compute the cost of a chain of
modules, in order to start the heavy long chains as early as possible in
order to avoid late long-tails. Through this we maximize concurrency
over the whole compilation period.

We use the file size as the cost / weight value, which is a rather poor
proxy for actual compilation time but it's cheap to get.

It's hard to talk about measurable improvements. There are several test
projects that are improved by this new algorithm but it's hard to prove
on a M4 Pro MacBook as it has too many CPU cores and so most projects
are maximally concurrent regardless of how you prioritize across
modules. Artifically reducing concurrency to --jobs 2 does show how this
branch typically results in a speedup. The implementation has a trivial
cost, so there are likely no negative scenarios - none have been
witnessed during the development of this patch.