Add BuildSpec & actonc spec subcommands
This adds a new BuildSpec module, which is the internal data
representation for a "build spec", that is the configuration we have for
an Acton project, currently in build.act.json, including dependencies
and zig dependencies.

There is a new actonc spec command for working with the build
spec (deliberately not build-spec to avoid tab completion overlap with
the build command), offering two subcommands: spec dump to dump the
build spec from Build.act (or build.act.json as fallback) in JSON format
— useful for tools that want the build spec as JSON — and spec update
FILE which can update Build.act in place with updates from a JSON file
such as build.act.json, to enable programmatic interaction with the
build spec for tools that do not understand .act source code. It does a
reasonable attempt to touch as little as possible, just replacing
dependencies & zig_dependencies, though keys may be reordered.