Bump syntax version to rebuild after int change
We're abusing the syntax version (we haven't really changed the syntax)
to indicate whenever we need a rebuild to happen. In this case we've
changed out the int type (in a previous branch / PR), so all code needs
to be considered stale and recompile. Bumping version is the easy way to
get that!