Parse the most negative int
-9223372036854775808 is the most negative values of our int (i64). The
standard lexer produces two tokens: - 9223372036854775808, which means
this is constructed as an int64 type with the value 9223372036854775808
prefixed by the unary operator -. However, 9223372036854775808 is too
large to be stored in a i64, so it fails.

We now parse negative integers into actual negative integers by
combining the UMinus with the int value in the parser (after the lexer).

Added a few tests and some helpers to parse just Stmt or Expr, this seems to
work.