Override conflicting dependencies
If we find multiple different versions of a dependency, we now override it
using the left-most, i.e. anything that we have found earlier which typically
means that dependencies listed in the main project or higher up dependencies
will override those of lower deps. This ensures a consistent build and
avoids various build errors. A warning is printed when a conflict is
overridden.

Matching on name is really too naive in order to match up actual
conflicts, but we don't have anything better for the time being and it's
frequent enough of a problem that this is worth it right now.