Bump GC size & harmonize LSP-server
We statically bump up the size for the GC nursery / generation 0 to 64MB
which yields quite a good performance improvement. We already did this
when invoking actonc from acton, but now it's done statically so direct
actonc calls are also faster.

Also, the LSP-server gets the same plus threads are enabled since we
want to start running tasks in threads or at the very least experiment
with it.