Dispatch `acton build` to `actonc build`
We now use the new total build graph in actonc as the default way to
build projects.

We also have `acton build --old-build` for the old builder - not really sure
why anyone would want that, but I guess we'll keep it for a bit before cleaning
it away.

The biggest change is probably that actonc properly overrides deps on
conflicts so that we get a consistent set of dependencies across the
build graph.