Apply dep overrides for declared deps
We mistakenly applied dep overrides unconditionally, thus introducing
dependencies in a project that didn't declare them. This also led to
compile cycles as a dependency could get a dependency to itself.

We now only override the path with --dep overrides for dependencies that
are actually declared in the project (Build.act or build.act.json).