Fix build of unused dep
This was rather interesting. In a project we can declare dependencies on
other projects. Since we now have a total build graph, we only compile
modules in dependencies that are actually imported from the main
project. What's interesting is that in some cases we depend on things
that are entirely unused. In this case, in the sorespo repo, in its gen
part, it depends on Orchestron which in turn depends on netconf. In the
normal main sorespo build, it uses netconf, but in the gen part, it
doesn't use any part of Orchestron that in turn uses netconf and so we
end up with no modules compiled in the netconf dep. Then we proceed with
the zig build, which then tries to build an archive out of the netconf
project but since there are no compiled modules, there are no .c files
and thus nothing to link or put together and that's a Zig error. We fix
this by adding a dummy file with a dummy function in case we didn't find
anything else.